+++ src/app/(auth)/layout.tsx
@@ -2,7 +2,6 @@ import type { Metadata } from "next";
 import { ClerkProvider } from "@clerk/nextjs";
 import { shadcn } from "@clerk/themes";
 import { Geist, Geist_Mono } from "next/font/google";
-import { SpeedInsights } from "@vercel/speed-insights/next";
 import ConvexClientProvider from "@/components/convex-client-provider";
 
 import { BaseTheme } from "@clerk/types";
@@ -56,7 +55,6 @@ export default function RootLayout({
             >
               {children}
               <ThemeChangeButton className="fixed bottom-6 left-6" />
-              <SpeedInsights />
             </ThemeProvider>
           </ConvexClientProvider>
         </ClerkProvider>